# add channels last support for MaxPool2d on CPU path