--- a/aten/src/ATen/native/DilatedMaxPool2d.cpp
+++ b/aten/src/ATen/native/DilatedMaxPool2d.cpp
@@ -1,9 +1,7 @@
 #include <ATen/ATen.h>
-#include <ATen/Parallel.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/NamedTensorUtils.h>
 #include <ATen/native/Pool.h>
-#include <tuple>
 
 
 namespace at {
@@ -11,117 +9,10 @@ namespace native {
 
 namespace {
 
-template <typename scalar_t>
-static void max_pool2d_with_indices_single_out_frame(
-          scalar_t *input_p,
-          scalar_t *output_p,
-          int64_t *ind_p,
-          int64_t nslices,
-          int64_t iwidth,
-          int64_t iheight,
-          int64_t owidth,
-          int64_t oheight,
-          int kW,
-          int kH,
-          int dW,
-          int dH,
-          int padW,
-          int padH,
-          int dilationW,
-          int dilationH
-          )
-{
-  at::parallel_for(0, nslices, 0, [&](int64_t start, int64_t end) {
-    for (auto k = start; k < end; k++)
-    {
-      /* loop over output */
-      int64_t i, j;
-      scalar_t *ip = input_p   + k*iwidth*iheight;
-      for(i = 0; i < oheight; i++)
-      {
-        for(j = 0; j < owidth; j++)
-        {
-          int64_t hstart = i * dH - padH;
-          int64_t wstart = j * dW - padW;
-          int64_t hend = std::min(hstart + (kH - 1) * dilationH + 1, iheight);
-          int64_t wend = std::min(wstart + (kW - 1) * dilationW + 1, iwidth);
-          while(hstart < 0)
-            hstart += dilationH;
-          while(wstart < 0)
-            wstart += dilationW;
-
-          /* local pointers */
-          scalar_t *op = output_p  + k*owidth*oheight + i*owidth + j;
-          int64_t *indp = ind_p   + k*owidth*oheight + i*owidth + j;
-
-          /* compute local max: */
-          int64_t maxindex = hstart*iwidth + wstart;
-          scalar_t maxval = -std::numeric_limits<scalar_t>::infinity();
-          for(int64_t y = hstart; y < hend; y += dilationH)
-          {
-            for(int64_t x = wstart; x < wend; x += dilationW)
-            {
-              int64_t tcntr = y*iwidth + x;
-              scalar_t val = *(ip + tcntr);
-              if ((val > maxval) || std::isnan(val))
-              {
-                maxval = val;
-                maxindex = tcntr;
-              }
-            }
-          }
-
-          /* set output to local max */
-          *op = maxval;
-
-          /* store location of max */
-          *indp = maxindex;
-        }
-      }
-    }
-  });
-}
-
-template <typename scalar_t>
-static void max_pool2d_with_indices_out_frame(
-          scalar_t *input_data,
-          scalar_t *output_data,
-          int64_t *indices_data,
-          int64_t nbatch,
-          int64_t nInputPlane,
-          int64_t inputWidth,
-          int64_t inputHeight,
-          int64_t outputWidth,
-          int64_t outputHeight,
-          int kW,
-          int kH,
-          int dW,
-          int dH,
-          int padW,
-          int padH,
-          int dilationW,
-          int dilationH)
-{
-  at::parallel_for(0, nbatch, 0, [&](int64_t start, int64_t end) {
-    for (auto p = start; p < end; p++) {
-      max_pool2d_with_indices_single_out_frame(
-        input_data+p*nInputPlane*inputWidth*inputHeight,
-        output_data+p*nInputPlane*outputWidth*outputHeight,
-        indices_data+p*nInputPlane*outputWidth*outputHeight,
-        nInputPlane,
-        inputWidth, inputHeight,
-        outputWidth, outputHeight,
-        kW, kH, dW, dH,
-        padW, padH,
-        dilationW, dilationH);
-    }
-  });
-}
-
 void max_pool2d_with_indices_out_cpu_template(
           Tensor& output,
           Tensor& indices,
-          const Tensor& input_,
+          const Tensor& input,
           IntArrayRef kernel_size,
           IntArrayRef stride,
           IntArrayRef padding,
@@ -152,152 +43,50 @@ void max_pool2d_with_indices_out_cpu_template(
   const int dilationH = safe_downcast<int, int64_t>(dilation[0]);
   const int dilationW = dilation.size() == 1 ? dilationH : safe_downcast<int, int64_t>(dilation[1]);
 
-  TORCH_CHECK((input_.ndimension() == 3 || input_.ndimension() == 4),
+  TORCH_CHECK((input.ndimension() == 3 || input.ndimension() == 4),
     "non-empty 3D or 4D (batch mode) tensor expected for input");
 
+  TORCH_CHECK(input.dtype() == output.dtype(),
+    "expected dtype ", input.dtype(), " for `output` but got dtype ", output.dtype());
+
   /* sizes */
-  const int64_t nbatch = input_.ndimension() == 4 ? input_.size(-4) : 1;
-  const int64_t nInputPlane = input_.size(-3);
-  const int64_t inputHeight = input_.size(-2);
-  const int64_t inputWidth = input_.size(-1);
+  const int64_t nbatch = input.ndimension() == 4 ? input.size(-4) : 1;
+  const int64_t nInputPlane = input.size(-3);
+  const int64_t inputHeight = input.size(-2);
+  const int64_t inputWidth = input.size(-1);
 
   const int64_t outputHeight = pooling_output_shape<int64_t>(inputHeight, kH, padH, dH, dilationH, ceil_mode);
   const int64_t outputWidth = pooling_output_shape<int64_t>(inputWidth, kW, padW, dW, dilationW, ceil_mode);
 
   pool2d_shape_check(
-    input_,
+    input,
     kH, kW, dH, dW, padH, padW, dilationH, dilationW,
     nInputPlane,
     inputHeight, inputWidth,
     outputHeight, outputWidth);
 
-  /* get contiguous input */
-  Tensor input = input_.contiguous();
-
-  /* resize output */
-  if (input.ndimension() == 3)
-  {
+  /* resize output and indices */
+  if (input.ndimension() == 3) {
     output.resize_({nInputPlane, outputHeight, outputWidth});
     /* indices will contain the locations for each output point */
     indices.resize_({nInputPlane, outputHeight, outputWidth});
-
-    AT_DISPATCH_FLOATING_TYPES(input.scalar_type(),
-      "max_pool2d_with_indices_cpu",
-      [&] {
-        /* get raw pointers */
-        scalar_t *input_data = input.data_ptr<scalar_t>();
-        scalar_t *output_data = output.data_ptr<scalar_t>();
-        int64_t *indices_data = indices.data_ptr<int64_t>();
-
-        max_pool2d_with_indices_single_out_frame(
-          input_data, output_data,
-          indices_data,
-          nInputPlane,
-          inputWidth, inputHeight,
-          outputWidth, outputHeight,
-          kW, kH, dW, dH,
-          padW, padH,
-          dilationW, dilationH);
-      }
-    );
-  }
-  else
-  {
-    output.resize_({nbatch, nInputPlane, outputHeight, outputWidth});
+  } else {
+    output.resize_({nbatch, nInputPlane, outputHeight, outputWidth}, input.suggest_memory_format());
     /* indices will contain the locations for each output point */
-    indices.resize_({nbatch, nInputPlane, outputHeight, outputWidth});
-
-    AT_DISPATCH_FLOATING_TYPES(input.scalar_type(),
-      "max_pool2d_with_indices_cpu",
-      [&] {
-        scalar_t *input_data = input.data_ptr<scalar_t>();
-        scalar_t *output_data = output.data_ptr<scalar_t>();
-        int64_t *indices_data = indices.data_ptr<int64_t>();
-
-        max_pool2d_with_indices_out_frame(
-          input_data,
-          output_data,
-          indices_data,
-          nbatch,
-          nInputPlane,
-          inputWidth, inputHeight,
-          outputWidth, outputHeight,
-          kW, kH, dW, dH,
-          padW, padH,
-          dilationW, dilationH); }
-    );
+    indices.resize_({nbatch, nInputPlane, outputHeight, outputWidth}, input.suggest_memory_format());
   }
-}
 
-template <typename scalar_t>
-static void max_pool2d_with_indices_backward_single_out_frame(
-          scalar_t *gradInput_p,
-          scalar_t *gradOutput_p,
-          int64_t *ind_p,
-          int64_t nInputPlane,
-          int64_t inputWidth,
-          int64_t inputHeight,
-          int64_t outputWidth,
-          int64_t outputHeight,
-          int dW,
-          int dH)
-{
-  at::parallel_for(0, nInputPlane, 0, [&](int64_t start, int64_t end) {
-    for (auto k = start; k < end; k++)
-    {
-      scalar_t *gradInput_p_k = gradInput_p + k*inputWidth*inputHeight;
-      scalar_t *gradOutput_p_k = gradOutput_p + k*outputWidth*outputHeight;
-      int64_t *ind_p_k = ind_p + k*outputWidth*outputHeight;
-
-      /* calculate max points */
-      int64_t i, j;
-      for(i = 0; i < outputHeight; i++)
-      {
-        for(j = 0; j < outputWidth; j++)
-        {
-          /* retrieve position of max */
-          int64_t maxp = ind_p_k[i*outputWidth + j];
-          if (maxp != -1) {
-            /* update gradient */
-            gradInput_p_k[maxp] += gradOutput_p_k[i*outputWidth + j];
-          }
-        }
-      }
-    }
-  });
-}
-
-template <typename scalar_t>
-static void max_pool2d_with_indices_backward_out_frame(
-          scalar_t *gradInput_data,
-          scalar_t *gradOutput_data,
-          int64_t *indices_data,
-          int64_t nbatch,
-          int64_t nInputPlane,
-          int64_t inputWidth,
-          int64_t inputHeight,
-          int64_t outputWidth,
-          int64_t outputHeight,
-          int dW,
-          int dH)
-{
-  at::parallel_for(0, nbatch, 0, [&](int64_t start, int64_t end) {
-    for (auto p = start; p < end; p++) {
-      max_pool2d_with_indices_backward_single_out_frame<scalar_t>(
-        gradInput_data+p*nInputPlane*inputWidth*inputHeight,
-        gradOutput_data+p*nInputPlane*outputWidth*outputHeight,
-        indices_data+p*nInputPlane*outputWidth*outputHeight,
-        nInputPlane,
-        inputWidth, inputHeight,
-        outputWidth, outputHeight,
-        dW, dH);
-    }
-  });
+  max_pool2d_kernel(
+      kCPU, output, indices, input,
+      kW, kH,
+      dW, dH,
+      padW, padH,
+      dilationW, dilationH);
 }
 
 Tensor& max_pool2d_with_indices_backward_out_cpu_template(
           Tensor& gradInput,
-          const Tensor& gradOutput_,
+          const Tensor& gradOutput,
           const Tensor& input,
           const Tensor& indices,
           IntArrayRef kernel_size,
@@ -333,11 +122,13 @@ Tensor& max_pool2d_with_indices_backward_out_cpu_template(
   TORCH_CHECK((input.ndimension() == 3 || input.ndimension() == 4),
     "non-empty 3D or 4D (batch mode) tensor expected for input");
 
-  /* get contiguous gradOutput */
-  const Tensor gradOutput = gradOutput_.contiguous();
+  TORCH_CHECK(input.dtype() == gradOutput.dtype(),
+    "expected dtype ", input.dtype(), " for `gradOutput` but got dtype ", gradOutput.dtype());
+  TORCH_CHECK(input.dtype() == gradInput.dtype(),
+    "expected dtype ", input.dtype(), " for `gradInput` but got dtype ", gradInput.dtype());
 
   /* resize */
-  gradInput.resize_as_(input);
+  gradInput.resize_(input.sizes(), input.suggest_memory_format());
   gradInput.zero_();
 
   /* sizes */
@@ -354,7 +145,7 @@ Tensor& max_pool2d_with_indices_backward_out_cpu_template(
 
   max_pool2d_backward_shape_check(
     input,
-    gradOutput_,
+    gradOutput,
     indices,
     nbatch,
     kH, kW, dH, dW, padH, padW, dilationH, dilationW,
@@ -362,48 +153,7 @@ Tensor& max_pool2d_with_indices_backward_out_cpu_template(
     inputHeight, inputWidth,
     outputHeight_for_shape_check, outputWidth_for_shape_check);
 
-  /* backprop */
-  if (input.ndimension() == 3)
-  {
-    AT_DISPATCH_FLOATING_TYPES(input.scalar_type(),
-      "max_pool2d_with_indices_backward",
-      [&] {
-        /* get raw pointers */
-        scalar_t *gradInput_data = gradInput.data_ptr<scalar_t>();
-        scalar_t *gradOutput_data = gradOutput.data_ptr<scalar_t>();
-        int64_t *indices_data = indices.data_ptr<int64_t>();
-
-        max_pool2d_with_indices_backward_single_out_frame(
-          gradInput_data, gradOutput_data,
-          indices_data,
-          nInputPlane,
-          inputWidth, inputHeight,
-          outputWidth, outputHeight,
-          dW, dH);
-      }
-    );
-  }
-  else
-  {
-    AT_DISPATCH_FLOATING_TYPES(input.scalar_type(),
-      "max_pool2d_with_indices_backward",
-      [&] {
-        /* get raw pointers */
-        scalar_t *gradInput_data = gradInput.data_ptr<scalar_t>();
-        scalar_t *gradOutput_data = gradOutput.data_ptr<scalar_t>();
-        int64_t *indices_data = indices.data_ptr<int64_t>();
-
-        max_pool2d_with_indices_backward_out_frame<scalar_t>(
-          gradInput_data, gradOutput_data,
-          indices_data,
-          nbatch,
-          nInputPlane,
-          inputWidth, inputHeight,
-          outputWidth, outputHeight,
-          dW, dH);
-      }
-    );
-  }
+  max_pool2d_backward_kernel(kCPU, gradInput, gradOutput, indices);
 
   return gradInput;
 }
@@ -463,7 +213,7 @@ std::tuple<Tensor, Tensor> max_pool2d_with_indices_cpu(
 
 Tensor& max_pool2d_with_indices_backward_out_cpu(
   Tensor& gradInput,
-  const Tensor& gradOutput_,
+  const Tensor& gradOutput,
   const Tensor& input,
   IntArrayRef kernel_size,
   IntArrayRef stride,
@@ -474,7 +224,7 @@ Tensor& max_pool2d_with_indices_backward_out_cpu(
 {
   max_pool2d_with_indices_backward_out_cpu_template(
     gradInput,
-    gradOutput_,
+    gradOutput,
     input,
     indices,
     kernel_size,
@@ -486,7 +236,7 @@ Tensor& max_pool2d_with_indices_backward_out_cpu(
 }
 
 Tensor max_pool2d_with_indices_backward_cpu(
-  const Tensor& gradOutput_,
+  const Tensor& gradOutput,
   const Tensor& input,
   IntArrayRef kernel_size,
   IntArrayRef stride,
@@ -495,10 +245,10 @@ Tensor max_pool2d_with_indices_backward_cpu(
   bool ceil_mode,
   const Tensor& indices)
 {
-  auto gradInput = at::zeros_like(input, LEGACY_CONTIGUOUS_MEMORY_FORMAT);
+  auto gradInput = at::empty({0}, input.options());
   max_pool2d_with_indices_backward_out_cpu_template(
     gradInput,
-    gradOutput_,
+    gradOutput,
     input,
     indices,
     kernel_size,
@@ -508,6 +258,9 @@ Tensor max_pool2d_with_indices_backward_cpu(
     ceil_mode);
   return gradInput;
 }
+
+DEFINE_DISPATCH(max_pool2d_kernel);
+DEFINE_DISPATCH(max_pool2d_backward_kernel);
 
 } // at::native
 } // at