… aten/src/ATen/native/Pool.h
… aten/src/ATen/native/cpu/AdaptiveAvgPoolKernel.cpp
… aten/src/ATen/native/cpu/MaxPoolKernel.cpp
… aten/src/ATen/native/cpu/UpSampleKernel.cpp
… aten/src/ATen/native/cpu/utils.h
… test/test_nn.py