This patch adds channels last support for `nn.MaxPool2d`.