@mingfeima 2020-08-07
This patch adds channels last memory format support for `nn.MaxPool2d` on CPU, CL path is manually vectorized with `vec256` on dimension of C.

Performance result on  CPU Intel(R) Xeon(R) Platinum 8260 CPU @ 2.40GHz, 2*20 cores, comparing CL v.s. contiguous:
* single core inference (BS=1): **9.6x** speedup
* single socket inference (BS=1): **6.6x** speedup
* single socket inference (BS=128): **4.0x** speedup

input size is picked up from RN50. Use this [script](https://github.com/mingfeima/op_bench-py/blob/master/max_pool2d.py) to reproduce, `./run.sh max_pool2d.py`.

```bash
### (before)
### using OMP_NUM_THREADS=1
MaxPool2d(contiguous): input_size [1, 64, 112, 112] time: 5.317 ms

### using OMP_NUM_THREADS=20
MaxPool2d(contiguous): input_size [1, 64, 112, 112] time: 6.443 ms
MaxPool2d(contiguous): input_size [128, 64, 112, 112] time: 45.178 ms

### (after)
### using OMP_NUM_THREADS=1
MaxPool2d(contiguous):   input_size [1, 64, 112, 112] time: 5.304 ms
MaxPool2d(channels_last): input_size [1,   64, 112, 112] time: 0.555 ms

### using OMP_NUM_THREADS=20
MaxPool2d(contiguous):   input_size [1, 64, 112, 112] time: 0.336 ms
MaxPool2d(channels_last): input_size [1,   64, 112, 112] time: 0.051 ms
MaxPool2d(contiguous): input_size [128,   64, 112, 112] time: 42.141 ms
MaxPool2d(channels_last): input_size   [128, 64, 112, 112] time: 10.455 ms
```