@mingfeima 2020-11-05
Updates:
* stack this patch on #42104
* add arg data type check for both forward and backward
* add device `CPU` for test case `test_max_pool2d_nhwc` from test_nn.py